+++ .github/workflows/algolia.yml
@@ -18,9 +18,10 @@ jobs:
         uses: actions/checkout@v3
       - name: Wait for Netlify deployment
         if: github.event_name != 'workflow_dispatch'
-        uses: probablyup/wait-for-netlify-action@3.2.0
+        uses: probablyup/wait-for-netlify-action@v3.4.0
         with:
           site_id: ${{ secrets.NETLIFY_SITE_ID }}
+          context: production
         env:
           NETLIFY_TOKEN: ${{ secrets.NETLIFY_TOKEN }}
       - name: Trigger Algolia crawl
