+++ package.json
@@ -60,7 +60,7 @@
     "url-loader": "4.1.1",
     "workbox-routing": "6.5.4",
     "workbox-strategies": "6.5.4",
-    "yup": "1.1.1"
+    "yup": "1.2.0"
   },
   "devDependencies": {
     "@commitlint/cli": "17.6.3",
